feat: squash recent staging commits
