--- webapp/alembic/versions/3328567a4fd1_added_the_author_table.py
@@ -1,41 +1,0 @@
-"""added the author table
-
-Revision ID: 3328567a4fd1
-Revises: 62f9c9a26cce
-Create Date: 2024-10-08 15:20:43.050216
-
-"""
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = "3328567a4fd1"
-down_revision = "62f9c9a26cce"
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        "author",
-        sa.Column("first_name", sa.String(), nullable=False),
-        sa.Column("last_name", sa.String(), nullable=False),
-        sa.Column("email", sa.String(), nullable=False),
-        sa.Column("asset", sa.Integer(), nullable=False),
-        sa.ForeignKeyConstraint(
-            ["asset"],
-            ["asset.id"],
-        ),
-        sa.PrimaryKeyConstraint("first_name", "last_name", "email"),
-        sa.UniqueConstraint("email"),
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table("author")
-    # ### end Alembic commands ###
