--- webapp/alembic/versions/aaf824b40866_added_model_for_author.py
@@ -1,33 +1,0 @@
-"""added model for author
-
-Revision ID: aaf824b40866
-Revises: 0113956b1765
-Create Date: 2024-10-08 16:12:23.032520
-
-"""
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = "aaf824b40866"
-down_revision = "0113956b1765"
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column(
-        "asset", sa.Column("author_email", sa.String(), nullable=True)
-    )
-    op.create_foreign_key(None, "asset", "author", ["author_email"], ["email"])
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(None, "asset", type_="foreignkey")
-    op.drop_column("asset", "author_email")
-    # ### end Alembic commands ###
